// src/app/sharing/Models/StudentDTO.ts
export interface StudentDTO {
  id: number;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  gender: string;
  email: string;
  phoneNumber: string;
  street: undefined,
  city: undefined,
  state: undefined,
  postalCode: undefined,
  country: undefined,
  enrollmentDate: String;
  isActive: boolean;
}